// src/app/model/bettingtip.ts
export class Bettingtip {
  id: string;
  event: string;
  tipp: string;
  odds: number;
  dateString: string;
  deadline: string;
  sportcategory: string;
  category: string;
  picUrl: string;
  datestamp: Date;
  analysis: string;
  tippster: string;
  tippUid: string;
  stake: number;
  result: string;
  isWin: boolean;
}